Address PR review: side-effect-free ADC key module, verify manual deploy refs
- ADC_CONTEXT_ENV_KEYS moves to client/src/lib/adcContextEnvKeys.ts so the
  credential-free test helper no longer evaluates firebaseAdmin.ts (which
  initializes Firebase Admin at module scope) just to read the constant;
  firebaseAdmin.ts re-exports it to keep one source of truth.
- deploy.yml workflow_dispatch now resolves the requested ref to a SHA and
  fails closed unless it is an ancestor of main WITH a successful CI run for
  that exact commit — manual roll-forward can no longer ship an arbitrary or
  red SHA.
