+++ client/src/lib/firebaseAdmin.ts
@@ -40,20 +40,10 @@ const firebaseConfigForAdmin = {
     process.env.FIREBASE_STORAGE_BUCKET || "blueprint-8c1ca.appspot.com",
 };
 
-/**
- * Every env key that can make this module attempt credential loading or an
- * Application Default Credentials lookup. Credential-free test harnesses must
- * strip ALL of these from subprocess environments — clearing only the first
- * two leaves GOOGLE_CLOUD_PROJECT/K_SERVICE able to trigger an ADC attempt.
- */
-export const ADC_CONTEXT_ENV_KEYS = Object.freeze([
-  "FIREBASE_SERVICE_ACCOUNT_JSON",
-  "GOOGLE_APPLICATION_CREDENTIALS",
-  "GOOGLE_CLOUD_PROJECT",
-  "GCLOUD_PROJECT",
-  "K_SERVICE",
-  "FUNCTION_TARGET",
-] as const);
+// The list of env keys that trigger credential loading / ADC lookups below
+// lives in the side-effect-free adcContextEnvKeys.ts module so test harnesses
+// can consume it without evaluating this module's admin initialization.
+export { ADC_CONTEXT_ENV_KEYS } from "./adcContextEnvKeys";
 
 function initializeFirebaseAdmin() {
   if (!admin) {
